+++ core/src/main/java/org/neo4j/ogm/metadata/FieldsInfo.java
@@ -13,12 +13,64 @@
 
 package org.neo4j.ogm.metadata;
 
+import java.lang.annotation.Annotation;
+import java.lang.reflect.*;
 import java.util.*;
+
+import org.neo4j.ogm.annotation.Transient;
 
 /**
  * @author Vince Bickers
  */
 public class FieldsInfo {
+
+    public static FieldsInfo create(Class<?> cls) {
+        Map<String, FieldInfo> fields = new HashMap<>();
+
+        for (Field field : cls.getDeclaredFields()) {
+            final int modifiers = field.getModifiers();
+            if (!Modifier.isTransient(modifiers) && !Modifier.isFinal(modifiers) && !Modifier.isStatic(modifiers)) {
+                ObjectAnnotations objectAnnotations = new ObjectAnnotations();
+                final Annotation[] declaredAnnotations = field.getDeclaredAnnotations();
+                for (Annotation annotation : declaredAnnotations) {
+                    AnnotationInfo info = AnnotationInfo.create(annotation);
+                    objectAnnotations.put(info.getName(), info);
+                }
+                if (objectAnnotations.get(Transient.class) == null) {
+                    String typeParameterDescriptor = null;
+                    final Type genericType = field.getGenericType();
+                    if (genericType instanceof ParameterizedType) {
+                        ParameterizedType parameterizedType = (ParameterizedType)genericType;
+                        final Type[] actualTypeArguments = parameterizedType.getActualTypeArguments();
+                        if (actualTypeArguments.length > 0) {
+                            for (Type typeArgument: actualTypeArguments) {
+                                if (typeArgument instanceof ParameterizedType) {
+                                    ParameterizedType parameterizedTypeArgument = (ParameterizedType)typeArgument;
+                                    typeParameterDescriptor = parameterizedTypeArgument.getRawType().getTypeName();
+                                    break;
+                                }
+                                else if (typeArgument instanceof TypeVariable || typeArgument instanceof WildcardType) {
+                                    typeParameterDescriptor = Object.class.getName();
+                                    break;
+                                }
+                                else if (typeArgument instanceof Class) {
+                                    typeParameterDescriptor = ((Class)typeArgument).getName();
+                                }
+                            }
+                        }
+                        if (typeParameterDescriptor == null) {
+                            typeParameterDescriptor = parameterizedType.getRawType().getTypeName();
+                        }
+                    }
+                    if (typeParameterDescriptor == null && (genericType instanceof TypeVariable)) {
+                        typeParameterDescriptor = field.getType().getTypeName();
+                    }
+                    fields.put(field.getName(), new FieldInfo(field.getName(), field.getType().getTypeName(),typeParameterDescriptor, objectAnnotations));
+                }
+            }
+        }
+        return new FieldsInfo(fields);
+    }
 
     private final Map<String, FieldInfo> fields;
