+++ core/src/main/java/org/neo4j/ogm/metadata/MetaData.java
@@ -25,7 +25,6 @@ import java.util.Set;
 import org.neo4j.ogm.annotation.NodeEntity;
 import org.neo4j.ogm.annotation.RelationshipEntity;
 import org.neo4j.ogm.exception.AmbiguousBaseClassException;
-import org.neo4j.ogm.metadata.builder.DomainInfoBuilder;
 import org.neo4j.ogm.typeconversion.ConversionCallback;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -43,7 +42,7 @@ public class MetaData {
     private Map<String, ClassInfo> classInfos = new HashMap<>();
 
     public MetaData(String... packages) {
-        domainInfo = DomainInfoBuilder.create(packages);
+        domainInfo = DomainInfo.create(packages);
     }
 
     /**
