+++ core/src/main/java/org/neo4j/ogm/metadata/MethodInfo.java
@@ -14,7 +14,11 @@
 package org.neo4j.ogm.metadata;
 
 
+import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
+import java.util.HashMap;
+import java.util.Map;
 
 import org.reflections.ReflectionUtils;
