+++ core/src/test/java/org/neo4j/ogm/annotation/relationshipEntities/AnnotatedFieldNonAnnotatedSetterAndNonSetterTest.java
@@ -26,7 +26,6 @@ import org.neo4j.ogm.annotation.EndNode;
 import org.neo4j.ogm.annotation.Relationship;
 import org.neo4j.ogm.annotation.RelationshipEntity;
 import org.neo4j.ogm.annotation.StartNode;
-import org.neo4j.ogm.metadata.builder.DomainInfoBuilder;
 import org.neo4j.ogm.metadata.reflect.EntityAccessManager;
 import org.neo4j.ogm.metadata.reflect.FieldWriter;
 import org.neo4j.ogm.metadata.reflect.RelationalWriter;
@@ -39,7 +38,7 @@ import org.neo4j.ogm.metadata.DomainInfo;
  */
 public class AnnotatedFieldNonAnnotatedSetterAndNonSetterTest {
     private EntityAccessManager entityAccessStrategy = new EntityAccessManager();
-    private DomainInfo domainInfo = DomainInfoBuilder.create("org.neo4j.ogm.annotation.relationshipEntities");
+    private DomainInfo domainInfo = DomainInfo.create("org.neo4j.ogm.annotation.relationshipEntities");
 
 
     @Test
