+++ core/src/test/java/org/neo4j/ogm/annotation/relationships/IterableRelationalReaderWriterTest.java
@@ -25,7 +25,6 @@ import org.junit.Before;
 import org.junit.Test;
 import org.neo4j.ogm.annotation.Relationship;
 import org.neo4j.ogm.domain.entityMapping.iterables.*;
-import org.neo4j.ogm.metadata.builder.DomainInfoBuilder;
 import org.neo4j.ogm.metadata.reflect.EntityAccessManager;
 import org.neo4j.ogm.metadata.reflect.FieldReader;
 import org.neo4j.ogm.metadata.reflect.FieldWriter;
@@ -47,7 +46,7 @@ public class IterableRelationalReaderWriterTest {
     @Before
     public void setup() {
         entityAccessStrategy = new EntityAccessManager();
-        domainInfo = DomainInfoBuilder.create("org.neo4j.ogm.domain.entityMapping.iterables");
+        domainInfo = DomainInfo.create("org.neo4j.ogm.domain.entityMapping.iterables");
     }
 
     /**
